+++ src/Api/Api.php
@@ -9,12 +9,11 @@
 
 namespace Joomla\Webservices\Api;
 
-use Joomla\Input\Input;
-use Joomla\Utilities\ArrayHelper;
 use Joomla\DI\Container;
 use Joomla\Webservices\Application;
 use Joomla\Event\DispatcherAwareInterface;
 use Joomla\Event\DispatcherInterface;
+use Joomla\Language\Text;
 
 /**
  * Interface to handle api calls
@@ -23,7 +22,7 @@ use Joomla\Event\DispatcherInterface;
  * @subpackage  Api
  * @since       1.2
  */
-class Api extends ApiBase implements DispatcherAwareInterface
+abstract class Api extends ApiBase implements DispatcherAwareInterface
 {
 	/**
 	 * @var    string  Name of the Api
@@ -54,6 +53,14 @@ class Api extends ApiBase implements DispatcherAwareInterface
 	protected $dispatcher = null;
 
 	/**
+	 * The text translation object
+	 *
+	 * @var    Text
+	 * @since  __DELPOY_VERSION__
+	 */
+	protected $text = null;
+
+	/**
 	 * Method to instantiate the file-based api call.
 	 *
 	 * @param   Container  $container  The DIC object
@@ -64,6 +71,7 @@ class Api extends ApiBase implements DispatcherAwareInterface
 	public function __construct(Container $container, $options = null)
 	{
 		$this->app = $container->get('app');
+		$this->text = $container->get('Joomla\\Language\\LanguageFactory')->getText();
 
 		// Initialise / Load options
 		$this->setOptions($options);
@@ -108,45 +116,6 @@ class Api extends ApiBase implements DispatcherAwareInterface
 	}
 
 	/**
-	 * Execute the Api operation.
-	 *
-	 * @return  mixed  JApi object with information on success, boolean false on failure.
-	 *
-	 * @since   1.2
-	 * @throws  \RuntimeException
-	 */
-	public function execute()
-	{
-		return null;
-	}
-
-	/**
-	 * Method to render the api call output.
-	 *
-	 * @return  string  Api call output
-	 *
-	 * @since   1.2
-	 */
-	public function render()
-	{
-		return '';
-	}
-
-	/**
-	 * Change the debug mode
-	 *
-	 * @param   boolean  $debug  Enable / Disable debug
-	 *
-	 * @return  void
-	 *
-	 * @since   1.2
-	 */
-	public function setDebug($debug)
-	{
-		$this->options->set('debug', (boolean) $debug);
-	}
-
-	/**
 	 * Load extension language file.
 	 *
 	 * @param   string  $option  Option name
@@ -166,63 +135,5 @@ class Api extends ApiBase implements DispatcherAwareInterface
 		|| $lang->load($option, $path . "/components/$option", $lang->getDefault(), false, false);
 
 		return $this;
-	}
-
-	/**
-	 * Returns posted data in array format
-	 *
-	 * @param   Container  $container  The DIC object
-	 *
-	 * @return  array
-	 *
-	 * @since   1.2
-	 */
-	public static function getPostedData(Container $container)
-	{
-		/** @var \Joomla\Input\Input $input */
-		$input = $container->get('app')->input;
-		$inputData = file_get_contents("php://input");
-
-		if (is_object($inputData))
-		{
-			$inputData = ArrayHelper::fromObject($inputData);
-		}
-		elseif(is_string($inputData))
-		{
-			$parsedData = null;
-
-			// We try to transform it into JSON
-			if ($data_json = @json_decode($inputData, true))
-			{
-				if (json_last_error() == JSON_ERROR_NONE)
-				{
-					$parsedData = (array) $data_json;
-				}
-			}
-
-			// We try to transform it into XML
-			if (is_null($parsedData) && $xml = @simplexml_load_string($inputData))
-			{
-				$json = json_encode((array) $xml);
-				$parsedData = json_decode($json, true);
-			}
-
-			// We try to transform it into Array
-			if (is_null($parsedData) && !empty($inputData) && !is_array($inputData))
-			{
-				parse_str($inputData, $parsedData);
-			}
-
-			$inputData = $parsedData;
-		}
-		else
-		{
-			$inputData = $input->post->getArray();
-		}
-
-		// Filter data with Input default filter
-		$postedData = new Input($inputData);
-
-		return $postedData->getArray();
 	}
 }
